Fixing a bug in jitter related to exception handling
Note:
Let T be the set of programs runnable with the previous version and W the set of program which run now then
I can say for sure that : T is subset of W

Advanteges:
Any program which run in previous version will use the same execution path in the new version.
Some program which fail in previous version will run. Now I can run call annotation's properties.

However, my solution is not complete because:

1 - Verification errors can emerge
2 - W is a subset of U, the set of valid programs

Final solution to this is required.